feat: 优化阅读记录，增加精确的最后阅读时间戳并改进进度保存逻辑
使用 `lastRead` 字段来记录精确的阅读时间，并优化了视频、音频及漫画模块在退出或暂停时的进度保存触发时机。

+++ app/src/main/java/io/legado/app/data/entities/ReadRecordShow.kt
@@ -1,10 +1,10 @@
 package io.legado.app.data.entities
 
 /**
- * UI 展示用聚合数据：lastRead 为本地日期 yyyyMMdd 整数键。
+ * UI 展示用聚合数据：lastRead 为当日最后一次阅读的毫秒时间戳，0 表示历史数据缺失。
  */
 data class ReadRecordShow(
     var bookName: String,
     var readTime: Long,
-    var lastRead: Int
+    var lastRead: Long
 )
